rag-app/backend/api/middleware.py
"""
Basic security headers for the FastAPI application
"""

from starlette.middleware.base import BaseHTTPMiddleware


class SecurityHeadersMiddleware(BaseHTTPMiddleware):
    """Add basic security headers to all responses"""

    async def dispatch(self, request, call_next):
        response = await call_next(request)

        # Basic security headers
        response.headers["X-Content-Type-Options"] = "nosniff"
        response.headers["X-Frame-Options"] = "DENY"

        return response
